// src/api-client/src/api-client/models/userVO.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { TokenVO } from './tokenVO'

export interface UserVO {
  /** 主键，雪花ID前端需注意精度问题！ */
  id?: string
  /** 登录账号 */
  username?: string
  /** 昵称 */
  nickname?: string
  /** 头像地址 */
  avatarUrl?: string
  /** 性别:0-未知,1-男,2-女 */
  gender?: string
  /** 用户类型:1-管理员,2-普通用户 */
  userType?: string
  /** 状态:0-禁用,1-启用 */
  status?: string
  /** 逻辑删除标记:0-否,1-是 */
  isDeleted?: string
  /** 最后登录时间 */
  lastLoginTime?: string
  /** 用户所有权限Code */
  permissionCodes?: string[]
  /** 用户所有角色Code */
  roleCodes?: string[]
  /** token,前端需要保存其中的tokenName和tokenValue */
  token?: TokenVO
}